requirements.txt
boto3==1.35.79
botocore==1.35.79
dlib==19.24.6
fastapi==0.115.6
jose==1.0.0
numpy==1.23.4
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
passlib==1.7.4
Pillow==11.1.0
pydantic[email]==2.10.5
python-dotenv==1.0.1
python_jose==3.3.0
scipy==1.11.3
SQLAlchemy==2.0.36
starlette==0.41.0
uvicorn==0.34.0
psycopg2-binary==2.9.8
python-multipart==0.0.19